added dynamic routes with ssr on headlines and news
