+++ ReadMe.md
@@ -4,3 +4,18 @@
 
 | ![Thumbnail 1](/thumbnail.png) | ![Thumbnail 2](/portrait-thumbnail.png) |
 | :----------------------------: | :-------------------------------------: |
+
+## Progress
+
+-   [x] **Home Page**
+    -   [x] **Weather**
+    -   [x] **Subscribed News**
+    -   [x] **Promoted News**
+-   [x] **Headlines Page**
+    -   [x] **Server Side Rendering**
+-   [ ] **Events Page**
+    -   work in progress
+-   [ ] **Lost Found Page**
+    -   work in progress
+-   [ ] **Profile Page**
+    -   almost done, edit profile page left
